LayoutTests/platform/gtk/accessibility/text-in-span-block-in-a-block-expected.txt

one
before
two
three
before,
four
,after
before
five
,after
before,
six
This tests that text inside "a" block inside "span" block is accessible.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS accText is "AXValue:one"
PASS setCaretOffset(elem, 1) is true
PASS accText is "AXValue:two"
PASS setCaretOffset(elem, 1) is true
PASS accText is "AXValue:three"
PASS setCaretOffset(elem, 1) is true
PASS accText is "AXValue:before,four,after"
PASS setCaretOffset(elem, 1) is true
PASS accText is "AXValue:five,after"
PASS setCaretOffset(elem, 1) is true
PASS accText is "AXValue:before,six"
PASS setCaretOffset(elem, 1) is true
PASS successfullyParsed is true

TEST COMPLETE
